Fixed memory leak and module
Memory leak was caused by malloc'ing for the translations file name in table.c,
I modified the way it is allocated (now done whether the default file is used or
not, so that it will need to be freed either way), and added a free.

There was a foolish error in the module, if the confirmation was received from the
keyboard, an error would be printed and the message would be sent. The error obviously
shouldn't be printed there, so I moved that statement to the 'else' condition, to be
output when the confirmation is not received.